requirements.txt
certifi==2020.6.20
cffi==1.14.2
configparser==4.0.2
cryptography==2.8
oci==2.21.1
pycparser==2.20
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
XlsxWriter==1.3.3
